refactor(error): centralize runtime error contract
