+++ lib/jido_ai/directive/llm_generate.ex
@@ -78,9 +78,8 @@ defimpl Jido.AgentServer.DirectiveExec, for: Jido.AI.Directive.LLMGenerate do
   when an agent is created.
   """
 
-  alias Jido.AI.{Observe, Signal, Turn}
+  alias Jido.AI.{Error, Observe, Signal, Turn}
   alias Jido.AI.Directive.Helpers
-  alias Jido.AI.Signal.Helpers, as: SignalHelpers
 
   def exec(directive, _input_signal, state) do
     %{
@@ -174,7 +173,7 @@ defimpl Jido.AgentServer.DirectiveExec, for: Jido.AI.Directive.LLMGenerate do
            signal =
              Signal.LLMResponse.new!(%{
                call_id: call_id,
-               result: SignalHelpers.normalize_result(result, :llm_error, "LLM request failed"),
+               result: Error.normalize_result(result, :llm_error, "LLM request failed"),
                metadata: signal_metadata(event_meta)
              })
 
@@ -189,7 +188,7 @@ defimpl Jido.AgentServer.DirectiveExec, for: Jido.AI.Directive.LLMGenerate do
             call_id: call_id,
             result:
               {:error,
-               SignalHelpers.error_envelope(
+               Error.error_envelope(
                  :supervisor,
                  "Failed to start LLM task",
                  %{reason: inspect(reason)},
