+++ lib/jido_ai/reasoning/react/runner.ex
@@ -12,7 +12,7 @@ defmodule Jido.AI.Reasoning.ReAct.Runner do
   alias Jido.AI.Reasoning.ReAct.{Config, Event, PendingToolCall, State, Token, ToolSelection}
   alias Jido.AI.Effects
   alias Jido.AI.Context, as: AIContext
-  alias Jido.AI.Signal.Helpers, as: SignalHelpers
+  alias Jido.AI.Error
   alias Jido.AI.Turn
   alias Jido.Agent.State, as: AgentState
 
@@ -906,7 +906,7 @@ defmodule Jido.AI.Reasoning.ReAct.Runner do
     max_retries = normalize_retry_count(config.tool_exec[:max_retries])
     backoff_ms = normalize_backoff(config.tool_exec[:retry_backoff_ms])
 
-    case SignalHelpers.retryable?(result) and attempt <= max_retries do
+    case Error.retryable?(result) and attempt <= max_retries do
       true ->
         case backoff_ms > 0 do
           true -> Process.sleep(backoff_ms)
